In Mapper, the '\t' characters in input line (SAM records) should be relaced with space sice '\t' is used for key-value separation.
In Reducer, it is sufficient to use the Java class IdentityReducer and no reducer.py is needed.
Since the IdentityReducer takes the value part of the inputs and prints them unchanged, the Mapper output should be set in the value part.
Therefore, the key-value pair in the Mapper should be :'something'+'\t'+SAM.line. 'something' is used as key for the reducer and will be ignored.
----
.gitignore: The *.class and *.jar files should not be tracked by git.